create task definitions interfaces
